update (SMAPI 3.18.6, SDV 1.5.6)

+++ source/Modular Gameplay Overhaul/ModApi.cs
@@ -44,7 +44,7 @@ public sealed class ModApi
 {
     #region professions
 
-    /// <summary>Get the value of an Ecologist's forage quality.</summary>
+    /// <summary>Gets the value of an Ecologist's forage quality.</summary>
     /// <param name="farmer">The player.</param>
     /// <returns>A <see cref="SObject"/> quality level.</returns>
     public int GetEcologistForageQuality(Farmer? farmer = null)
@@ -53,7 +53,7 @@ public sealed class ModApi
         return farmer.HasProfession(Profession.Ecologist) ? farmer.GetEcologistForageQuality() : SObject.lowQuality;
     }
 
-    /// <summary>Get the value of a Gemologist's mineral quality.</summary>
+    /// <summary>Gets the value of a Gemologist's mineral quality.</summary>
     /// <param name="farmer">The player.</param>
     /// <returns>A <see cref="SObject"/> quality level.</returns>
     public int GetGemologistMineralQuality(Farmer? farmer = null)
@@ -62,7 +62,7 @@ public sealed class ModApi
         return farmer.HasProfession(Profession.Gemologist) ? farmer.GetGemologistMineralQuality() : SObject.lowQuality;
     }
 
-    /// <summary>The price bonus applied to animal produce sold by Producer.</summary>
+    /// <summary>Gets the price bonus applied to animal produce sold by Producer.</summary>
     /// <param name="farmer">The player.</param>
     /// <returns>A <see cref="float"/> multiplier for animal products.</returns>
     public float GetProducerProducePriceBonus(Farmer? farmer = null)
@@ -71,7 +71,7 @@ public sealed class ModApi
         return farmer.GetProducerPriceBonus();
     }
 
-    /// <summary>The price bonus applied to fish sold by Angler.</summary>
+    /// <summary>Gets the price bonus applied to fish sold by Angler.</summary>
     /// <param name="farmer">The player.</param>
     /// <returns>A <see cref="float"/> multiplier for fish prices.</returns>
     public float GetAnglerFishPriceBonus(Farmer? farmer = null)
@@ -81,7 +81,7 @@ public sealed class ModApi
     }
 
     /// <summary>
-    ///     Get the value of the a Conservationist's effective tax deduction based on the preceding season's trash
+    ///     Gets the value of the a Conservationist's effective tax deduction based on the preceding season's trash
     ///     collection.
     /// </summary>
     /// <param name="farmer">The player.</param>
@@ -469,7 +469,7 @@ public sealed class ModApi
     /// <param name="monster">The <see cref="Monster"/>.</param>
     /// <param name="duration">The duration in milliseconds.</param>
     /// <param name="intensity">The intensity of the slow effect.</param>
-    public void Slow(Monster monster, int duration, double intensity = 0.5)
+    public void Slow(Monster monster, int duration, float intensity = 0.5f)
     {
         monster.Slow(duration, intensity);
     }
